Final 4.4.7 bug fixes

+++ components/scifio-tools/src/loci/formats/tools/ImageInfo.java
@@ -926,10 +926,14 @@ public class ImageInfo {
       LOGGER.info("Generating OME-XML (schema version {})", version);
     }
     if (ms instanceof MetadataRetrieve) {
-      service.removeBinData(service.getOMEMetadata((MetadataRetrieve) ms));
-      for (int i=0; i<reader.getSeriesCount(); i++) {
-        service.addMetadataOnly(
-          service.getOMEMetadata((MetadataRetrieve) ms), i);
+      // adding MetadataOnly elements to an OME-TIFF's XML will cause
+      // validation errors
+      if (!(baseReader instanceof OMETiffReader)) {
+        service.removeBinData(service.getOMEMetadata((MetadataRetrieve) ms));
+        for (int i=0; i<reader.getSeriesCount(); i++) {
+          service.addMetadataOnly(
+            service.getOMEMetadata((MetadataRetrieve) ms), i);
+        }
       }
 
       if (omexmlOnly) {
